streamsight_complete/backend/simulator/requirements.txt
kafka-python==2.0.2
pymongo==4.6.1
python-dotenv==1.0.0
